telegram/requirements.txt
paho-mqtt
requests
python-telegram-bot
python-dotenv
